Feat/1/add database support

+++ server/src/usecase/usecase.rs
@@ -1,19 +1,19 @@
-use crate::domain::{ClientName, Repository, Task, ThreeNResult};
+use crate::domain::{ClientName, RawTask, Repository, ThreeNResult};
 
-fn allocate_new_task<R: Repository>(repo: &mut R) -> Task {
+fn allocate_new_task<R: Repository>(repo: &mut R) -> RawTask {
     let state = repo.fetch_current_state();
     let from = state.from;
     let to = state.from + state.increment - 1;
 
     repo.update_from(state.from + state.increment);
-    Task { from, to }
+    RawTask { from, to }
 }
 
-pub fn handle_ready<R: Repository>(repo: &mut R, client_name: String) -> Task {
+pub fn handle_ready<R: Repository>(repo: &mut R, client_name: String) -> RawTask {
     // check if this client has an existing tasks, if so return it again
     // otherwise, allocate new task
     if let Some(task) = repo.fetch_queued_task_by_client(client_name.clone()) {
-        return task;
+        return task.to_raw_task();
     }
 
     let task = allocate_new_task(repo);
@@ -25,19 +25,25 @@ pub fn handle_ready<R: Repository>(repo: &mut R, client_name: String) -> Task {
 pub fn handle_solved<R: Repository>(
     repo: &mut R,
     client_name: ClientName,
-    task: Task,
+    task: RawTask,
     result: ThreeNResult,
 ) -> () {
-    if repo
-        .fetch_queued_task_by_client(client_name.clone())
-        .is_none()
-    {
-        eprintln!("Unknown problem");
-        return;
+    match repo.fetch_queued_task_by_client(client_name.clone()) {
+        None => {
+            eprintln!("Unknown problem");
+        }
+        Some(task_timestamp) => {
+            if task.from != task_timestamp.from || task.to != task_timestamp.to {
+                eprintln!(
+                    "Task range doesn't match! Got {:?} but queued was {:?}",
+                    task, task_timestamp
+                );
+                return;
+            }
+            repo.delete_queued_task_by_client(client_name.clone());
+            repo.store_results(client_name, task_timestamp, result);
+        }
     }
-
-    repo.delete_queued_task_by_client(client_name.clone());
-    repo.store_results(client_name, task, result);
 }
 
 #[cfg(test)]
@@ -60,7 +66,7 @@ mod tests {
 
         assert_eq!(
             task,
-            Task {
+            RawTask {
                 from: 1,
                 to: 1_000_000
             }
@@ -100,6 +106,23 @@ mod tests {
     }
 
     #[test]
+    fn when_solving_task_provided_must_match_queued() {
+        let mut repo = InMemoryRepository::new();
+        let task = handle_ready(&mut repo, client_name());
+        let wrong_task = RawTask { from: 1, to: 1 };
+        handle_solved(&mut repo, client_name(), wrong_task, result());
+
+        let solved = repo.fetch_results_by_client(client_name());
+        let task_queued_for_client = repo.fetch_queued_task_by_client(client_name());
+
+        assert_eq!(solved.len(), 0);
+        assert!(task_queued_for_client.is_some());
+        let task_queued = task_queued_for_client.unwrap();
+        assert_eq!(task_queued.from, task.from);
+        assert_eq!(task_queued.to, task.to);
+    }
+
+    #[test]
     fn after_ready_after_solve_works() {
         let mut repo = InMemoryRepository::new();
         let task = handle_ready(&mut repo, client_name());
@@ -109,7 +132,7 @@ mod tests {
 
         assert_eq!(
             task2,
-            Task {
+            RawTask {
                 from: 1_000_001,
                 to: 2_000_000
             }
